Bug 1465232 - Separate Top Tabs displaying into a new TabDisplayManager (#3959)
* Bug 1465232 - Separate Top Tabs displaying into a new TabDisplayManager

* fix indentation
